+++ EducationZone-Backend/src/main/java/com/example/EducationZoneBackend/Service/StudentService.java
@@ -7,6 +7,7 @@ import com.example.EducationZoneBackend.Exceptions.NotFoundException;
 import com.example.EducationZoneBackend.Models.Student;
 import com.example.EducationZoneBackend.Repository.StudentRepository;
 import lombok.SneakyThrows;
+import org.dozer.DozerBeanMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -22,94 +23,87 @@ public class StudentService {
     @Autowired
     public StudentService(StudentRepository studentRepository, KeycloakAdminService keycloakAdminService) {
         this.studentRepository = studentRepository;
-
         this.keycloakAdminService = keycloakAdminService;
     }
 
 
-
     @SneakyThrows
-    public void registerStudent(RegisterStudentDTO registerStudentDto)
-    {
-        System.out.println(registerStudentDto);
-
-        if(studentRepository.findByUsername(registerStudentDto.getUsername()).isPresent())
-        {
+    public void registerStudent(RegisterStudentDTO registerStudentDto) {
+        if (studentRepository.findByUsername(registerStudentDto.getUsername()).isPresent()) {
             throw new AlreadyExistException("Username Already Exist");
         }
 
-        if(studentRepository.findByEmail(registerStudentDto.getEmail()).isPresent())
-        {
+        if (studentRepository.findByEmail(registerStudentDto.getEmail()).isPresent()) {
             throw new AlreadyExistException("Email Already Exist");
         }
 
-        Student student= Student.builder()
+        Student student = Student.builder()
                 .firstName(registerStudentDto.getFirstName())
                 .lastName(registerStudentDto.getLastName())
                 .email(registerStudentDto.getEmail())
                 .password(registerStudentDto.getPassword())
                 .username(registerStudentDto.getUsername())
-                .studentSet(registerStudentDto.getStudentSet())
-                .year(registerStudentDto.getYear())
-                .studentGroup(registerStudentDto.getStudentGroup())
+                .groupNumber(registerStudentDto.getGroupNumber())
                 .phone(registerStudentDto.getPhone())
+                .year(registerStudentDto.getYear())
+                .department(registerStudentDto.getDepartment())
                 .build();
 
         studentRepository.save(student);
-       // keycloakAdminService.registerUser(registerStudentDto.getUsername(), registerStudentDto.getPassword(), "ROLE_STUDENT");
-
+        //keycloakAdminService.registerUser(registerStudentDto.getUsername(), registerStudentDto.getPassword(), "ROLE_STUDENT");
 
     }
 
 
     @SneakyThrows
-    public void removeStudent(Long studentId)
-    {
-        Student student = studentRepository.findById(studentId).orElseThrow(()->new NotFoundException("student not found"));
-
+    public void removeStudent(Long studentId) {
+        Student student = studentRepository.findById(studentId).orElseThrow(() -> new NotFoundException("student not found"));
         studentRepository.delete(student);
 
     }
 
     @SneakyThrows
-    public GetStudentDTO getStudent(Long studentId)
-    {
-        Student student = studentRepository.findById(studentId).orElseThrow(()->new NotFoundException("student not found"));
+    public GetStudentDTO getStudent(Long studentId) {
+        Student student = studentRepository.findById(studentId).orElseThrow(() -> new NotFoundException("student not found"));
 
-        GetStudentDTO getStudentDto= GetStudentDTO.builder()
-                .firstName(student.getFirstName())
-                .lastName(student.getLastName())
-                .email(student.getEmail()).build();
+        //Dest dest = mapper.map(source, Dest.class);
+        GetStudentDTO getStudentDto = new DozerBeanMapper().map(student, GetStudentDTO.class);
 
         return getStudentDto;
     }
 
     @SneakyThrows
-    public void putStudent(Long studentId, RegisterStudentDTO newRegisterStudentDto)
-    {
-        Student student = studentRepository.findById(studentId).orElseThrow(()->new NotFoundException("student not found"));
+    public void updateStudent(Long studentId, RegisterStudentDTO newRegisterStudentDto) {
+        Student student = studentRepository.findById(studentId).orElseThrow(() -> new NotFoundException("student not found"));
 
-        //daca este admin sau este chiar studentul
+        if (newRegisterStudentDto.getFirstName() != null)
+            student.setFirstName(newRegisterStudentDto.getFirstName());
 
+        if (newRegisterStudentDto.getLastName() != null)
+            student.setLastName(newRegisterStudentDto.getLastName());
 
-            if (newRegisterStudentDto.getFirstName() != null)
-                student.setFirstName(newRegisterStudentDto.getFirstName());
+        if (newRegisterStudentDto.getEmail() != null)
+            student.setEmail(newRegisterStudentDto.getEmail());
 
-            if (newRegisterStudentDto.getLastName() != null)
-                student.setLastName(newRegisterStudentDto.getLastName());
+        if (newRegisterStudentDto.getPassword() != null)
+            student.setPassword(newRegisterStudentDto.getPassword());
 
-            if (newRegisterStudentDto.getEmail() != null)
-                student.setEmail(newRegisterStudentDto.getEmail());
+        if (newRegisterStudentDto.getUsername() != null)
+            student.setUsername(newRegisterStudentDto.getUsername());
 
-            if (newRegisterStudentDto.getPassword() != null)
-                student.setPassword(newRegisterStudentDto.getPassword());
+        if (newRegisterStudentDto.getGroupNumber() != null)
+            student.setGroupNumber(newRegisterStudentDto.getGroupNumber());
 
-            if (newRegisterStudentDto.getUsername() != null)
-                student.setUsername(newRegisterStudentDto.getUsername());
+        if (newRegisterStudentDto.getPhone() != null)
+            student.setPhone(newRegisterStudentDto.getPhone());
 
+        if (newRegisterStudentDto.getYear() != null)
+            student.setYear(newRegisterStudentDto.getYear());
 
-            studentRepository.save(student);
+        if (newRegisterStudentDto.getDepartment() != null)
+            student.setDepartment(newRegisterStudentDto.getDepartment());
 
+        studentRepository.save(student);
 
     }
 
@@ -117,7 +111,7 @@ public class StudentService {
     @SneakyThrows
     public List<GetStudentDTO> getAllStudents() {
 
-        if(studentRepository.findAllStudents().isEmpty())
+        if (studentRepository.findAllStudents().isEmpty())
             throw new NotFoundException("there are no students to display");
 
         return studentRepository.findAllStudents();
@@ -126,16 +120,19 @@ public class StudentService {
     @SneakyThrows
     public List<GetStudentDTO> getAllStudentsByName(String studentName) {
 
-
-        if(studentRepository.findAllStudentsByName(studentName).isEmpty())
-        {
+        if (studentRepository.findAllStudentsByName(studentName).isEmpty())
             throw new NotFoundException("Students Not Found");
-        }
+
 
         return studentRepository.findAllStudentsByName(studentName);
     }
 
+    @SneakyThrows
     public void removeAllStudents() {
+
+        if (studentRepository.findAllStudents().isEmpty())
+            throw new NotFoundException("Students Not Found");
+
         studentRepository.deleteAll();
     }
 }
